Support arbitrary clouds
We have a few more clouds on the way, and a lot of development clouds,
and I'd like to be able to tell internal users to use
terraform-provider-zpa, but they have to be able to point it at those
internal clouds.

This change will allow users to specify a precise URL instead of "",
"gov" or "beta", and have that go straight through.